requirements.txt
discord.py
python-dotenv
python-Levenshtein
